Head_body_merge
The above is my project contains two program files . the app.py is used to get the org img remove bg convert to cartoonized form crop the head then join to desired body. The another code file is used to take a cartoonized head and get the cartoonized body then merge it as a single processed image file.